Ensure HOME environmental variable set
When working with puppet (and other setups), it is possible the HOME environmental variable is unset. This
leads to a spf13 bootstrap which does not perform as expected because `$HOME` is empty. This results in files being written to undesirable locations such as `/.spf13-vim-3/.vimrc.bundles` Let's stop the user before they get into this bad state.